Storage and Firestore Service creation
